chore: remove release validator and isolate benchmark instrumentation
