+++ .gitattributes
@@ -3,10 +3,3 @@ docs-site/content/docs/reference/functions/** linguist-generated=true
 
 # Generated metadata consumed by docs components
 docs-site/src/generated/functions-meta.json linguist-generated=true
-
-# Unified diffs encode empty context lines as one space.
-benchmarks/*-baseline.patch whitespace=-blank-at-eol
-# This experimental diff also ends on an encoded empty context line.
-benchmarks/perf-ranges-422-a-only.patch whitespace=-blank-at-eol,-blank-at-eof
-# This patch ends with an encoded blank context line.
-benchmarks/perf-recalc-427-baseline.patch whitespace=-blank-at-eol,-blank-at-eof
